requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
APScheduler==3.10.4
python-dateutil==2.9.0
matplotlib==3.8.4
pytz==2024.1
